Program working and 7 tests

+++ StregSystemCore/IStregSystem.cs
@@ -10,7 +10,7 @@ namespace StregSystem.Core
     public interface IStregSystem
     {
         IEnumerable<Product> ActiveProducts { get; }
-        InsertCashTransaction AddCreditsToAccount(User user, int amount);
+        InsertCashTransaction AddCreditsToAccount(User user, decimal amount);
         BuyTransaction BuyProduct(User user, Product product);
         Product GetProductByID(int id);
         IEnumerable<Transaction> GetTransactions(User user, int count);
